move RQB folder
