+++ src/components/inputs/react-query-builder/combinator-selector.tsx
@@ -7,7 +7,7 @@
 
 import { CombinatorSelectorProps } from 'react-querybuilder';
 import { FunctionComponent, useCallback, useState } from 'react';
-import PopupConfirmationDialog from '../dialogs/popup-confirmation-dialog';
+import PopupConfirmationDialog from '../../dialogs/popup-confirmation-dialog.tsx';
 import { MaterialValueSelector } from '@react-querybuilder/material';
 
 const CombinatorSelector: FunctionComponent<CombinatorSelectorProps> = (
