+++ src/components/inputs/react-query-builder/custom-react-query-builder.tsx
@@ -25,14 +25,14 @@ import ValueEditor from './value-editor.tsx';
 import ValueSelector from './value-selector.tsx';
 import { useCallback, useMemo } from 'react';
 
-import { COMBINATOR_OPTIONS } from '../filter/expert/expert-filter-constants.ts';
-import ErrorInput from '../inputs/react-hook-form/error-management/error-input.tsx';
-import FieldErrorAlert from '../inputs/react-hook-form/error-management/field-error-alert.tsx';
+import { COMBINATOR_OPTIONS } from '../../filter/expert/expert-filter-constants.ts';
+import ErrorInput from '../react-hook-form/error-management/error-input.tsx';
+import FieldErrorAlert from '../react-hook-form/error-management/field-error-alert.tsx';
 import {
     countRules,
     getOperators,
     queryValidator,
-} from '../filter/expert/expert-filter-utils.ts';
+} from '../../filter/expert/expert-filter-utils.ts';
 
 interface CustomReactQueryBuilderProps {
     name: string;
